Binh - Fix some FE bugs

+++ BE/src/main/java/com/code/BE/repository/ConfirmationTokenRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ConfirmationTokenRepository extends JpaRepository<ConfirmationToken, Long> {
+public interface ConfirmationTokenRepository extends JpaRepository<ConfirmationToken, Integer> {
     ConfirmationToken findByConfirmationToken (String confirmationToken);
     ConfirmationToken findByUserId (int id);
     ConfirmationToken findById (int id);
